ログイン時バックエンドに渡すユーザー情報APIをコンポーネント化、それに伴いHomepage.tsxを修正
