+++ src/components/auth/RegisterForm.tsx
@@ -1,5 +1,5 @@
 import React from 'react';
-import { useForm } from '../../hooks/UseForm';
+import { useForm } from '../../hooks/useForm';
 import { validateRegister } from '../../utils/Validation';
 import { useNavigate } from 'react-router-dom';
 import { RegisterValues } from '../../interface/auth';
