+++ src/components/layout/Footer.tsx
@@ -1,4 +1,4 @@
-import '../css/footer.css';
+import '../../css/footer.css';
 import React from 'react';
 
 const Footer: React.FC = () => {
