+++ src/css/footer.css
@@ -1,4 +1,3 @@
-
 .footer {
   background-color: #333;
   color: white;
